requirements.txt
PyQt4==4.11.4
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.1